Mention Database Replication
Later in the document, you mention HyperDB, but really don't cover
it until that point. Also, HyperDB doesn't have to be the only way
to get Database Replication or Clustering. You should at least
mention the concept earlier. Your diagram has it as well, so might
be nice to round things out here.